pyproject.toml
[build-system]
# build the package with [flit](https://flit.readthedocs.io)
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
# See https://www.python.org/dev/peps/pep-0621/
name = "AITW_darcy_infiltration"
dynamic = ["version"]
description = "TODO"
authors = []
readme = "README.md"
license = { file = "LICENSE.txt" }
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python",
]
keywords = ["birch", "wood", "microstructure"]
requires-python = ">=3.10"
dependencies = [
    "rich_click",
    "tqdm",
    "numpy",

    "mpi4py",
    "petsc4py",
    "fenics-ufl",
    "fenics-basix",
]

# [project.urls]
# Source = ""

[project.optional-dependencies]
pre-commit = [
    "pre-commit",
]
release = [
    "flit"
]

[project.scripts]
aitw-darcy = "aitw_darcy.cli:cmd_root"

[tool.flit.module]
name = "aitw_darcy"

[tool.flit.sdist]
exclude = [
    ".gitignore", ".github", ".pre-commit-config.yaml",
]
